requirements.txt
asgiref==3.8.1
btcpay-python==1.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1
ecdsa==0.19.0
idna==3.8
psycopg==3.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
style==1.1.0
typing_extensions==4.12.2
urllib3==2.2.2
